% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.sparsegl-methods.R
\name{coef.cv.sparsegl}
\alias{coef.cv.sparsegl}
\title{Get coefficients or make coefficient predictions from a \code{cv.sparsegl} object.}
\usage{
\method{coef}{cv.sparsegl}(object, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{fitted \code{\link{cv.sparsegl}} object.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the value \code{s="lambda.1se"} stored
on the CV \code{object}, it is the largest value of \code{lambda} such that
error is within 1 standard error of the minimum. Alternatively
\code{s="lambda.min"} can be used, it is the optimal value of \code{lambda}
that gives minimum cross validation error \code{cvm}. If \code{s} is
numeric, it is taken as the value(s) of \code{lambda} to be used.}

\item{\dots}{not used. Other arguments to predict.}
}
\value{
The coefficients at the requested values for \code{lambda}.
}
\description{
This function gets coefficients or makes coefficient predictions from a
cross-validated \code{sparsegl} model, using the stored \code{"sparsegl.fit"}
object, and the optimal value chosen for \code{lambda}.
}
\details{
This function makes it easier to use the results of cross-validation to get
coefficients or make coefficient predictions.
}
\seealso{
\code{\link{cv.sparsegl}}, and \code{\link{predict.cv.sparsegl}}
methods.
}
