% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_irwls.R
\name{get_eta}
\alias{get_eta}
\title{Helper function to get etas (linear predictions) on the original scale}
\usage{
get_eta(x, xs, beta, b0)
}
\description{
Helper function to get etas (linear predictions) on the original scale
}
