% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gglasso.R
\name{plot.gglasso}
\alias{plot.gglasso}
\title{Plot solution paths from a "gglasso" object}
\usage{
\method{plot}{gglasso}(x, group = FALSE, log.l = TRUE, ...)
}
\arguments{
\item{x}{fitted \code{\link{gglasso}} model}

\item{group}{what is on the Y-axis. Plot the norm of each group if
\code{TRUE}. Plot each coefficient if \code{FALSE}.}

\item{log.l}{what is on the X-axis. Plot against the log-lambda sequence if
\code{TRUE}. Plot against the lambda sequence if \code{FALSE}.}

\item{\dots}{other graphical parameters to plot}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{\link{gglasso}} object.
}
\details{
A coefficient profile plot is produced.
}
\examples{

# load gglasso library
library(gglasso)

# load data set
data(bardet)

# define group index
group <- rep(1:20,each=5)

# fit group lasso
m1 <- gglasso(x=bardet$x,y=bardet$y,group=group,loss="ls")

# make plots
par(mfrow=c(1,3))
plot(m1) # plots the coefficients against the log-lambda sequence 
plot(m1,group=TRUE) # plots group norm against the log-lambda sequence 
plot(m1,log.l=FALSE) # plots against the lambda sequence
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang <yi.yang6@mcgill.ca>
}
\keyword{models}
\keyword{regression}
