% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparsegl.R
\name{plot.sparsegl}
\alias{plot.sparsegl}
\title{Plot solution paths from a \code{sparsegl} object}
\usage{
\method{plot}{sparsegl}(
  x,
  y_axis = c("coef", "group"),
  x_axis = c("lambda", "penalty"),
  add_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{fitted \code{\link{sparsegl}} model}

\item{y_axis}{variable is on the y_axis. Either the coefficients (default)
or the group norm.}

\item{x_axis}{variable on the x-axis. Either the (log)-lambda
sequence (default) or value of the penalty. The penalty is scaled by its
maximum along the path.}

\item{add_legend}{show the legend. Often, with many groups/predictors, this
can become overwhelming.}

\item{\dots}{other graphical parameters to plot}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{\link{sparsegl}} object. The result is a \code{ggplot}. Additional user
modifications can be added as desired.
}
