% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{predict.gglasso}
\alias{predict.gglasso}
\title{make predictions from a "gglasso" object.}
\usage{
\method{predict}{gglasso}(object, newx, s = NULL, type = c("class", "link"), ...)
}
\arguments{
\item{object}{fitted \code{\link{gglasso}} model object.}

\item{newx}{matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{type}{type of prediction required: \itemize{ \item Type \code{"link"},
for regression it returns the fitted response; for classification it gives
the linear predictors.  \item Type \code{"class"}, only valid for
classification, it produces the predicted class label corresponding to the
maximum probability.}}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this functions predicts fitted values and
class labels from a fitted \code{\link{gglasso}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s}
is not in the lambda sequence used for fitting the model, the \code{predict}
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of predicted values from both left and
right \code{lambda} indices.
}
\examples{

# load gglasso library
library(gglasso)

# load data set
data(colon)

# define group index
group <- rep(1:20,each=5)

# fit group lasso
m1 <- gglasso(x=colon$x,y=colon$y,group=group,loss="logit")

# predicted class label at x[10,]
print(predict(m1,type="class",newx=colon$x[10,]))

# predicted linear predictors at x[1:5,]
print(predict(m1,type="link",newx=colon$x[1:5,]))

}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\seealso{
\code{\link{coef}} method
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang <yi.yang6@mcgill.ca>
}
\keyword{models}
\keyword{regression}
