% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-methods.R
\name{predict.sparsegl}
\alias{predict.sparsegl}
\title{Make predictions from a \code{sparsegl} object.}
\usage{
\method{predict}{sparsegl}(object, newx, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{sparsegl}} model object.}

\item{newx}{matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on type.
}
\description{
Similar to other predict methods, this functions predicts fitted values and
class labels from a fitted \code{\link[=sparsegl]{sparsegl()}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s}
is not in the lambda sequence used for fitting the model, the \code{predict}
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of predicted values from both left and
right \code{lambda} indices.
}
\seealso{
\code{\link{coef}} method
}
