% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-methods.R
\name{print.sparsegl}
\alias{print.sparsegl}
\title{Print a \code{sparsegl} object}
\usage{
\method{print}{sparsegl}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{\link{sparsegl}} object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\value{
a two-column matrix, the first columns is the number of nonzero
group counts and the second column is \code{Lambda}.
}
\description{
Print the nonzero group counts at each lambda along the sparsegl path.
}
\details{
Print the information about the nonzero group counts at each lambda step in
the \code{\link{sparsegl}} object. The result is a two-column matrix with
columns \code{Df} and \code{Lambda}. The \code{Df} column is the number of
the groups that have nonzero within-group coefficients, the \code{Lambda}
column is the the corresponding lambda.
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
