% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-estimation.R
\name{risk_estimate}
\alias{risk_estimate}
\title{Information criteria}
\usage{
risk_estimate(object, x, y, type = c("AIC", "BIC", "GCV"), approx_df = FALSE)
}
\arguments{
\item{object}{fitted object from a call to \code{\link[=sparsegl]{sparsegl()}}}

\item{x}{matrix of predictors, of dimension \eqn{n \times p}{n*p}; each row
is a vector of measurement and each column is a feature}

\item{y}{real-valued response variable.}

\item{type}{one of AIC, BIC, or GCV}

\item{approx_df}{the \code{df} component of a \code{sparsegl} object is an approximation
(albeit a fairly accurate one) to the actual degrees-of-freedom. However,
the exact value requires inverting a portion of \verb{X'X}. So this may take
some time.}
}
\value{
a vector of the same length as \code{object$lambda}
}
\description{
Information criteria
}
