% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgl_irwls.R
\name{sgl_irwls}
\alias{sgl_irwls}
\title{Fit a GLM with sparse group regularization for a path of lambda values}
\usage{
sgl_irwls(
  bn,
  bs,
  ix,
  iy,
  nobs,
  nvars,
  x,
  y,
  pf,
  pfl1,
  dfmax,
  pmax,
  nlam,
  flmin,
  ulam,
  eps,
  maxit,
  vnames,
  group,
  intr,
  asparse,
  standardize,
  lower_bnd,
  upper_bnd,
  weights,
  offset = NULL,
  family = gaussian(),
  trace_it = 0,
  warm = NULL
)
}
\description{
Fit a generalized linear model via penalized maximum likelihood for a path of
lambda values. Can deal with any GLM family.
}
\details{
This organization is based largely off \code{\link[stats:glm]{stats::glm.fit()}} with some extras
to handle the path.

Sometimes the sequence is truncated before \code{nlambda} values of lambda
have been used. This happens when \code{glmnet.path} detects that the
decrease in deviance is marginal (i.e. we are near a saturated fit).
}
