% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD:man/coef.sparsegl.Rd
% Please edit documentation in R/tools.R
\name{coef.sparsegl}
\alias{coef.sparsegl}
\title{get coefficients or make coefficient predictions from an "sparsegl" object.}
=======
% Please edit documentation in R/sparsegl-methods.R
\name{coef.sparsegl}
\alias{coef.sparsegl}
\title{get coefficients or make coefficient predictions from an "gglasso" object.}
>>>>>>> main:man/coef.gglasso.Rd
\usage{
\method{coef}{sparsegl}(object, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{sparsegl}} model object.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{\dots}{not used. Other arguments to predict.}
}
\value{
The coefficients at the requested values for \code{lambda}.
}
\description{
Computes the coefficients at the requested values for \code{lambda} from a
fitted \code{\link{sparsegl}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s}
is not in the lambda sequence used for fitting the model, the \code{coef}
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of coefficients from both left and right
\code{lambda} indices.
}
<<<<<<< HEAD:man/coef.sparsegl.Rd
=======
\examples{

# load sparsegl library
library(sparsegl)

# load data set
data(colon)

# define group index
group <- rep(1:20, each=5)

# fit group lasso
m1 <- sparsegl(x=colon$x, y=colon$y, group=group)

# the coefficients at lambda = 0.01 and 0.02
coef(m1, s=c(0.01, 0.02))

}
>>>>>>> main:man/coef.gglasso.Rd
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\seealso{
\code{\link{predict.sparsegl}} method
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
