% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.sparsegl}
\alias{cv.sparsegl}
\alias{cv.ls}
\title{Cross-validation for a \code{sparsegl} object.}
\usage{
cv.sparsegl(
  x,
  y,
  group,
  lambda = NULL,
  pred.loss = c("L2", "L1"),
  nfolds = 5,
  foldid,
  ...
)
}
\arguments{
\item{x}{Matrix of predictors, of dimension \eqn{n \times p}{n * p}; each row
is an observation vector.}

\item{y}{Response variable. This argument should be quantitative for
regression (least squares).}

\item{group}{A vector of consecutive integers describing the grouping of the
coefficients (see example below).}

\item{lambda}{An optional user-supplied lambda sequence; default is
\code{NULL}, and \code{\link{sparsegl}} chooses its own sequence.}

\item{pred.loss}{Loss to use for cross-validation error. Valid options are:
\itemize{ \item \code{"L1"} for regression, mean square
error used by least squares regression \code{loss = "ls"}, it measure the
deviation from the fitted mean to the response.  \item \code{"L2"} for
regression, mean absolute error used by least squares regression
\code{loss = "ls"}, it measure the deviation from the fitted mean to the
response.  }Default is \code{"L2"}.}

\item{nfolds}{Number of folds - default is 5. Although \code{nfolds} can be
as large as the sample size (leave-one-out CV), it is not recommended for
large datasets. Smallest value allowable is \code{nfolds = 3}.}

\item{foldid}{An optional vector of values between 1 and \code{nfolds}
identifying which fold each observation is in. If supplied, \code{nfolds} can
be missing.}

\item{\dots}{Other arguments that can be passed to sparsegl.}
}
\value{
An object of class \code{\link{cv.sparsegl}} is returned, which is a
list with the ingredients of the cross-validation fit.  \item{lambda}{The
values of \code{lambda} used in the fits.} \item{cvm}{The mean
cross-validated error - a vector of length \code{length(lambda)}.}
\item{cvsd}{Estimate of standard error of \code{cvm}.} \item{cvupper}{Upper
curve = \code{cvm + cvsd}.} \item{cvlower}{Lower curve = \code{cvm - cvsd}.}
\item{name}{A text string indicating type of measure (for plotting
purposes).} \item{sparsegl.fit}{A fitted \code{\link{sparsegl}} object for the
full data.} \item{lambda.min}{The optimal value of \code{lambda} that gives
minimum cross validation error \code{cvm}.} \item{lambda.1se}{The largest
value of \code{lambda} such that error is within 1 standard error of the
minimum.}
}
\description{
Does k-fold cross-validation for a fitted \code{\link{sparsegl}} object.
This function is modified based on the \code{cv}
function from the \code{\link{glmnet-package}} package.
}
\details{
The function runs \code{\link{sparsegl}} \code{nfolds}+1 times; the first to
get the \code{lambda} sequence, and then the remainder to compute the fit
with each of the folds omitted. The average error and standard deviation
over the folds are computed.
}
\examples{
n <- 100
p <- 20
X <- matrix(rnorm(n * p), nrow = n)
eps <- rnorm(n)
beta_star <- c(rep(5, 5), c(5, -5, 2, 0, 0), rep(-5, 5), rep(0, (p - 15)))
y <- X \%*\% beta_star + eps
groups <- rep(1:(p / 5), each = 5)
fit1 <- sparsegl(X, y, group = groups)
cv_fit <- cv.sparsegl(X, y, groups)
}
\seealso{
\code{\link{sparsegl}}, \code{\link{plot.cv.sparsegl}},
\code{\link{predict.cv.sparsegl}}, and \code{\link{coef.cv.sparsegl}} methods.
}
