% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cv.sparsegl.R
\name{plot.cv.sparsegl}
\alias{plot.cv.sparsegl}
\title{plot the cross-validation curve produced from a \code{cv.sparsegl} object}
\usage{
\method{plot}{cv.sparsegl}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{fitted \code{\link{cv.sparsegl}} object}

\item{sign.lambda}{either plot against \code{log(lambda)} (default) or its
negative if \code{sign.lambda=-1}.}

\item{\dots}{other graphical parameters to plot}
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation
curves, as a function of the \code{lambda} values used. This function is
modified based on the \code{plot.cv} function from the \code{glmnet}
package.
}
\details{
A plot is produced.
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr

Friedman, J., Hastie, T., and Tibshirani, R. (2010), ``Regularization paths
for generalized linear models via coordinate descent,'' \emph{Journal of
Statistical Software}, 33, 1.\cr \url{http://www.jstatsoft.org/v33/i01/}
}
\seealso{
\code{\link{cv.sparsegl}}.
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
