% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparsegl.R
\name{plot.sparsegl}
\alias{plot.sparsegl}
\title{Plot solution paths from a "sparsegl" object}
\usage{
\method{plot}{sparsegl}(
  x,
  grouped = FALSE,
  asparse = 0.05,
  y_axis = c("coef", "group_norm"),
  x_axis = c("log_lambda", "norm"),
  ...
)
}
\arguments{
\item{x}{fitted \code{\link{sparsegl}} model}

\item{grouped}{if applying the user-specified grouping of features. Default
is FALSE.}

\item{asparse}{the weight to put on the ell1 norm in sparse group lasso. Default
is 0.05}

\item{y_axis}{what is on the Y_axis. Plot \code{"coef"} the coefficients or
\code{"norm"} group norm. Default is \code{"coef"}.}

\item{x_axis}{what is on the X-axis. Plot against the \code{"log_lambda"} log-lambda
sequence or \code{"norm"} a scaled norm vector. Default is \code{"log_lambda"}.}

\item{\dots}{other graphical parameters to plot}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{\link{sparsegl}} object.
}
\details{
A coefficient profile plot is produced.
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
