% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparsegl.R
\name{plot.sparsegl}
\alias{plot.sparsegl}
\title{Plot solution paths from a "sparsegl" object}
\usage{
\method{plot}{sparsegl}(x, group = FALSE, log.l = TRUE, alpha = 0.05, ...)
}
\arguments{
\item{x}{fitted \code{\link{sparsegl}} model}

\item{group}{what is on the Y-axis. Plot the norm of each group if
\code{TRUE}. Plot each coefficient if \code{FALSE}.}

\item{log.l}{what is on the X-axis. Plot against the log-lambda sequence if
\code{TRUE}. Plot against the lambda sequence if \code{FALSE}.}

\item{\dots}{other graphical parameters to plot}
}
\description{
Produces a coefficient profile plot of the coefficient paths for a fitted
\code{\link{sparsegl}} object.
}
\details{
A coefficient profile plot is produced.
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
