% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-methods.R, R/tools.R
\name{predict.sparsegl}
\alias{predict.sparsegl}
\title{Make predictions from a "sparsegl" object.}
\usage{
\method{predict}{sparsegl}(object, newx, s = NULL, ...)

\method{predict}{sparsegl}(object, newx, s = NULL, ...)
}
\arguments{
\item{object}{fitted \code{\link{sparsegl}} model object.}

\item{newx}{matrix of new values for \code{x} at which predictions are to be
made. Must be a matrix.}

\item{s}{value(s) of the penalty parameter \code{lambda} at which
predictions are required. Default is the entire sequence used to create the
model.}

\item{\dots}{Not used. Other arguments to predict.}
}
\value{
The object returned depends on type.

The object returned depends on type.
}
\description{
Similar to other predict methods, this functions predicts fitted values and
class labels from a fitted \code{\link[=sparsegl]{sparsegl()}} object.

Similar to other predict methods, this functions predicts fitted values and
class labels from a fitted \code{\link{sparsegl}} object.
}
\details{
\code{s} is the new vector at which predictions are requested. If \code{s}
is not in the lambda sequence used for fitting the model, the \code{predict}
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of predicted values from both left and
right \code{lambda} indices.

\code{s} is the new vector at which predictions are requested. If \code{s}
is not in the lambda sequence used for fitting the model, the \code{predict}
function will use linear interpolation to make predictions. The new values
are interpolated using a fraction of predicted values from both left and
right \code{lambda} indices.
}
\references{
Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr

Yang, Y. and Zou, H. (2015), ``A Fast Unified Algorithm for
Computing Group-Lasso Penalized Learning Problems,'' \emph{Statistics and
Computing}. 25(6), 1129-1141.\cr BugReport:
\url{https://github.com/emeryyi/gglasso}\cr
}
\seealso{
\code{\link{coef}} method

\code{\link{coef}} method
}
\author{
Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}

Yi Yang and Hui Zou\cr Maintainer: Yi Yang \href{mailto:yi.yang6@mcgill.ca}{yi.yang6@mcgill.ca}
}
\keyword{models}
\keyword{regression}
