% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-methods.R
\name{print.sparsegl}
\alias{print.sparsegl}
\title{Print a \code{sparsegl} object}
\usage{
\method{print}{sparsegl}(x, digits = min(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{fitted \code{\link{sparsegl}} object}

\item{digits}{significant digits in printout}

\item{\dots}{additional print arguments}
}
\description{
Prints a few summaries of the fitted model.
}
