% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsegl-package.R
\docType{package}
\name{sparsegl-package}
\alias{sparsegl-package}
\title{sparsegl: Sparse Group Lasso}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Efficient implementation of sparse group lasso with optional bound constraints on the coefficients; see \doi{10.18637/jss.v110.i06}. It supports the use of a sparse design matrix as well as returning coefficient estimates in a sparse matrix. Furthermore, it correctly calculates the degrees of freedom to allow for information criteria rather than cross-validation with very large data. Finally, the interface to compiled code avoids unnecessary copies and allows for the use of long integers.
}
\references{
Liang, X., Cohen, A., Sólon Heinsfeld, A., Pestilli, F., and
McDonald, D.J. 2024.
\emph{sparsegl: An \code{R} Package for Estimating Sparse Group Lasso.}
Journal of Statistical Software, Vol. 110(6): 1–23.
\doi{10.18637/jss.v110.i06}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dajmcdon/sparsegl}
  \item \url{https://dajmcdon.github.io/sparsegl/}
  \item Report bugs at \url{https://github.com/dajmcdon/sparsegl/issues}
}

}
\author{
\strong{Maintainer}: Daniel J. McDonald \email{daniel@stat.ubc.ca}

Authors:
\itemize{
  \item Xiaoxuan Liang \email{xiaoxuan.liang@stat.ubc.ca}
  \item Anibal Solón Heinsfeld \email{anibalsolon@gmail.com}
  \item Aaron Cohen \email{cohenaa@indiana.edu}
}

Other contributors:
\itemize{
  \item Yi Yang \email{yi.yang6@mcgill.ca} [contributor]
  \item Hui Zou [contributor]
  \item Jerome Friedman [contributor]
  \item Trevor Hastie [contributor]
  \item Rob Tibshirani [contributor]
  \item Balasubramanian Narasimhan [contributor]
  \item Kenneth Tay [contributor]
  \item Noah Simon [contributor]
  \item Junyang Qian [contributor]
  \item James Yang [contributor]
}

}
\keyword{internal}
